adding more options to optimizer

+++ include/optimizer.h
@@ -57,7 +57,7 @@ public:
 
 class Optimizer{
 public:
-	virtual OptResult optimize(OptFunction &function, std::vector<double> initial_params, double tol, int maxeval, std::vector<double> lowerBounds, std::vector<double> upperBounds) = 0;
+	virtual OptResult optimize(OptFunction &function, std::vector<double> initial_params, double tol, long long int maxeval, std::vector<double> lowerBounds, std::vector<double> upperBounds) = 0;
 	virtual const std::string get_algorithm() const { return ""; }
 	virtual const bool isGradientBased() const { return false; }
 };
@@ -68,7 +68,7 @@ struct ExtraNLOptData {
 
 class NLOptimizer : public Optimizer {
 public:
-  OptResult optimize(OptFunction &function, std::vector<double> initial_params, double tol, int maxeval, std::vector<double> lowerBounds, std::vector<double> upperBounds) override;
+  OptResult optimize(OptFunction &function, std::vector<double> initial_params, double tol, long long int maxeval, std::vector<double> lowerBounds, std::vector<double> upperBounds) override;
   const bool isGradientBased() const override;
   virtual const std::string get_algorithm() const;
 };
